requirements.txt
aiofiles
azure-identity
azure-search-documents
chardet
fastapi
Jinja2
langchain-community
langchain-milvus
langchain-openai
PyPDF2
python-multipart
uvicorn
python-docx
motor
azure-cognitiveservices-speech
pydantic_settings
beautifulsoup4